requirements.txt
streamlit
google-generativeai
youtube-transcript-api
fpdf2